etl_waterlevel/streamlit_app/requirements.txt
streamlit
pandas
psycopg2-binary
matplotlib
plotly-express
folium
streamlit-folium